Enclose field names containing literals in quotes
When creating the text representation of a path, field names that contain lexer literals (like a period ".') should be enclosed in quotes (or brackets).

The proposed fix only adds quotes if needed (as opposed to always). This avoids changing current paths/behavior drastically since it seems no one has run into this so far.